requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==1.0.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
magic-filter==1.0.12
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==2.0.2
opencv-python==4.12.0.88
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pillow==11.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
python-dateutil==2.8.2
pyzmq==25.1.1
six==1.16.0
spyder-kernels==2.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.7.1
wcwidth==0.2.9
wurlitzer==3.0.3
yarl==1.9.2
zipp==3.17.0
